fix(workspaces): some renaming
