+++ packages/server/assets/workspacesCore/typedefs/workspaces.graphql
@@ -519,6 +519,15 @@ extend type User {
     @hasServerRole(role: SERVER_GUEST)
     @hasScope(scope: "workspace:read")
     @isOwner
+
+  """
+  The last-visited workspace for the given user
+  """
+  activeWorkspace: Workspace
+  """
+  Returns `true` if last visited project was "legacy" "personal project" outside of a workspace
+  """
+  isProjectsActive: Boolean
 }
 
 extend type Project {
@@ -632,4 +641,8 @@ input AdminUpdateWorkspacePlanInput {
 
 type AdminMutations {
   updateWorkspacePlan(input: AdminUpdateWorkspacePlanInput!): Boolean!
+}
+
+extend type ActiveUserMutations {
+  setActiveWorkspace(slug: String, isProjectsActive: Boolean): Boolean!
 }
