+++ packages/server/modules/core/dbSchema.ts
@@ -288,14 +288,7 @@ export const StreamFavorites = buildTableHelper('stream_favorites', [
 export const UsersMeta = buildMetaTableHelper(
   'users_meta',
   ['userId', 'key', 'value', 'createdAt', 'updatedAt'],
-  [
-    'isOnboardingFinished',
-    'foo',
-    'bar',
-    'onboardingStreamId',
-    'activeWorkspace',
-    'activeProject'
-  ],
+  ['isOnboardingFinished', 'onboardingStreamId', 'activeWorkspace', 'isProjectsActive'],
   'userId'
 )
