+++ packages/server/modules/core/repositories/users.ts
@@ -30,8 +30,6 @@ import {
   MarkOnboardingComplete,
   MarkUserAsVerified,
   SearchLimitedUsers,
-  SetUserActiveProject,
-  SetUserActiveWorkspace,
   StoreUser,
   StoreUserAcl,
   UpdateUser,
@@ -571,18 +569,4 @@ export const searchUsersFactory =
       users: res.users.map(removePrivateFields),
       cursor: res.cursor
     }
-  }
-
-export const setUserActiveWorkspaceFactory =
-  (deps: { db: Knex }): SetUserActiveWorkspace =>
-  async ({ userId, workspaceSlug }) => {
-    const meta = metaHelpers(Users, deps.db)
-    await meta.set(userId, 'activeWorkspace', workspaceSlug)
-  }
-
-export const setUserActiveProjectFactory =
-  (deps: { db: Knex }): SetUserActiveProject =>
-  async ({ userId, projectId }) => {
-    const meta = metaHelpers(Users, deps.db)
-    await meta.set(userId, 'activeProject', projectId)
   }
